[DUOS-1668] Added DuosHeaderOld to preserve the old version better. The explicit name (vs the previous commit) makes more sense? And flipping it so the changes happen in the old file makes the diff easier to analyze
